Ajustes na aplicação com scanner

+++ LinguagemDeProgramacao/data-cat-jar/src/main/java/data/cat/service/MedidasServices.java
@@ -32,7 +32,7 @@ public class MedidasServices {
     }
 
     public Double getProcessadorEmUso() {
-        Double usoProcessador = looca.getProcessador().getUso().doubleValue();
+        Double usoProcessador = looca.getProcessador().getUso();
         return usoProcessador;
     }
